+++ code/WndDesign/style/style_helper.h
@@ -145,13 +145,13 @@ public:
 		return border._width > 0 && border._color != color_transparent; 
 	}
 	const RoundedRectangle GetBorder(Size display_size) const { 
-		return RoundedRectangle(display_size, border._radius, border._width, border._color); 
+		return RoundedRectangle(display_size, border._radius, (float)border._width, border._color); 
 	}
 	bool IsPointInside(Size display_size, Point point) const {
 		return PointInRoundedRectangle(display_size, border._radius, point);
 	}
 	const Margin CalculateBorderMargin() const {
-		return { border._width, border._width, border._width, border._width };
+		return { (int)border._width, (int)border._width, (int)border._width, (int)border._width };
 	}
 	const Rect GetDisplayRegionWithoutBorder(Size display_size) const {
 		return ShrinkRegionByMargin(Rect(point_zero, display_size), CalculateBorderMargin());
